+++ src/NNFactoryImp.hpp
@@ -17,8 +17,8 @@ namespace geopm
 
             std::shared_ptr<LocalNeuralNet> createLocalNeuralNet(const std::vector<std::shared_ptr<DenseLayer>> &layers) const override;
             std::shared_ptr<DenseLayer> createDenseLayer(const TensorTwoD &weights, const TensorOneD &biases) const override;
-            TensorTwoD createTensorTwoD(const std::vector<std::vector<float>> &vals) const override;
-            TensorOneD createTensorOneD(const std::vector<float> &vals) const override;
+            TensorTwoD createTensorTwoD(const std::vector<std::vector<double>> &vals) const override;
+            TensorOneD createTensorOneD(const std::vector<double> &vals) const override;
     };
 }
